backjoon-exam13

+++ algorithm-test/app/src/main/java/algorithm/test/backjoon/level01/exam12/Main.java
@@ -8,9 +8,9 @@ public class Main {
   public static void main(String[] args) {
     Scanner sc = new Scanner(System.in);
 
-    long a = sc.nextInt();
-    long b = sc.nextInt();
-    long c = sc.nextInt();
+    long a = sc.nextLong();
+    long b = sc.nextLong();
+    long c = sc.nextLong();
 
     System.out.print(a + b + c);
   }
